requirements.txt
crewai==0.11.2
web3==7.11.1
alchemy-sdk==0.1.1
python-dotenv==1.1.0
langchain==0.1.20
openai==1.79.0
pydantic==2.11.4 